+++ tests/some_module.py
@@ -1,5 +1,40 @@
 """Bogus module for testing."""
-# pylint: disable=missing-docstring,disallowed-name,invalid-name
+# pylint: disable=missing-docstring,disallowed-name,invalid-name,no-self-use
+
+MODULE_LEVEL_ATTRIBUTE = "test"
+
+
+class SomeClass:
+    CLASS_VALUE = "class_method"
+
+    def __init__(self):
+        self.instance_value = "instance_method"
+
+    @classmethod
+    def class_method(cls):
+        return cls.CLASS_VALUE
+
+    @classmethod
+    def class_method_with_args(cls, a):
+        return a
+
+    @staticmethod
+    def static_method():
+        return "static_method"
+
+    @staticmethod
+    def static_method_with_args(a):
+        return a
+
+    def instance_method(self):
+        return self.instance_value
+
+    def instance_method_with_args(self, a):
+        return a
+
+
+class DerivedClass(SomeClass):
+    pass
 
 
 class ModuleClass:
