+++ tox.ini
@@ -29,12 +29,11 @@ whitelist_externals=
     sh
 
 commands =
-    python -m unittest tests/test_flexmock.py
     python tests/test_unittest.py
     python tests/test_doctest.py
-    python -c "from twisted.scripts.trial import run; run();" tests/test_flexmock.py
-    zope-testrunner  --test-path=./ --test-file-pattern=test_flexmock --verbose
+    python -c "from twisted.scripts.trial import run; run();" tests/test_integrations.py
+    zope-testrunner  --test-path=./ --test-file-pattern=test_integrations --verbose
     python -m testtools.run --verbose tests/test_testtools.py
-    sh -c 'python -m subunit.run tests/test_flexmock.py | subunit2pyunit'
+    sh -c 'python -m subunit.run tests/test_integrations.py | subunit2pyunit'
     python tests/test_teamcity.py
     pytest tests/test_pytest.py
